# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StaffFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  companyName: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  companyName: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Employee {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  role: string;
  /** @nullable */
  department?: string | null;
  salary: number;
  /** monthly or daily */
  salaryType: string;
  joiningDate: string;
  /** active or inactive */
  status: string;
  adminId: number;
  createdAt: string;
}

export interface CreateEmployeeBody {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  role: string;
  /** @nullable */
  department?: string | null;
  salary: number;
  salaryType: string;
  joiningDate: string;
}

export interface UpdateEmployeeBody {
  name?: string;
  phone?: string;
  /** @nullable */
  email?: string | null;
  role?: string;
  /** @nullable */
  department?: string | null;
  salary?: number;
  salaryType?: string;
  joiningDate?: string;
  status?: string;
}

export interface Attendance {
  id: number;
  employeeId: number;
  employeeName: string;
  date: string;
  /** present, absent, half_day */
  status: string;
  /** @nullable */
  checkIn?: string | null;
  /** @nullable */
  checkOut?: string | null;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  overtimeHours?: number | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  createdAt: string;
}

export type MarkAttendanceBodyRecordsItem = {
  employeeId: number;
  date: string;
  status: string;
  /** @nullable */
  checkIn?: string | null;
  /** @nullable */
  checkOut?: string | null;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  overtimeHours?: number | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
};

export interface MarkAttendanceBody {
  records: MarkAttendanceBodyRecordsItem[];
}

export interface UpdateAttendanceBody {
  status?: string;
  /** @nullable */
  checkIn?: string | null;
  /** @nullable */
  checkOut?: string | null;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  overtimeHours?: number | null;
}

export interface AttendanceSummary {
  employeeId: number;
  employeeName: string;
  present: number;
  absent: number;
  halfDay: number;
  overtime: number;
  workingDays: number;
  percentage: number;
}

export interface Payment {
  id: number;
  employeeId: number;
  employeeName: string;
  amount: number;
  /** salary, advance, bonus, deduction */
  type: string;
  /** cash, upi, bank */
  method: string;
  /** @nullable */
  month?: number | null;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  note?: string | null;
  /** paid, pending */
  status: string;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}

export interface CreatePaymentBody {
  employeeId: number;
  amount: number;
  type: string;
  method: string;
  /** @nullable */
  month?: number | null;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  note?: string | null;
  status: string;
}

export interface UpdatePaymentBody {
  amount?: number;
  type?: string;
  method?: string;
  /** @nullable */
  note?: string | null;
  status?: string;
}

export interface SalarySummary {
  employeeId: number;
  employeeName: string;
  baseSalary: number;
  salaryType: string;
  workingDays: number;
  presentDays: number;
  halfDays: number;
  overtimeHours: number;
  overtimePay: number;
  deductions: number;
  advances: number;
  netSalary: number;
  paid: number;
  pending: number;
}

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  employeeId: number;
  employeeName: string;
  /** pending, in_progress, done */
  status: string;
  /** low, medium, high */
  priority: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  proofUrl?: string | null;
  createdAt: string;
}

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  employeeId: number;
  priority: string;
  /** @nullable */
  dueDate?: string | null;
}

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: string;
  priority?: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  proofUrl?: string | null;
}

export interface DashboardStats {
  totalEmployees: number;
  presentToday: number;
  absentToday: number;
  halfDayToday: number;
  totalSalaryThisMonth: number;
  totalPaidThisMonth: number;
  pendingDues: number;
  activeTasks: number;
}

export interface Insight {
  id: string;
  /** info, warning, success, alert */
  type: string;
  message: string;
  /** @nullable */
  metric?: string | null;
}

export interface AttendanceTrendDay {
  date: string;
  present: number;
  absent: number;
  halfDay: number;
}

export interface MonthlyReport {
  month: number;
  year: number;
  totalEmployees: number;
  totalWorkingDays: number;
  totalSalaryPaid: number;
  totalPending: number;
  attendanceRate: number;
  employees: SalarySummary[];
}

export type ListEmployeesParams = {
  search?: string;
  role?: string;
};

export type ListAttendanceParams = {
  employeeId?: number;
  month?: number;
  year?: number;
  date?: string;
};

export type GetAttendanceSummaryParams = {
  month: number;
  year: number;
  employeeId?: number;
};

export type ListPaymentsParams = {
  employeeId?: number;
  month?: number;
  year?: number;
  type?: string;
};

export type GetSalarySummaryParams = {
  month: number;
  year: number;
};

export type ListTasksParams = {
  employeeId?: number;
  status?: string;
};

export type GetMonthlyReportParams = {
  month: number;
  year: number;
};