# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StaffFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance for one or more employees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkAttendanceBody"
      responses:
        "200":
          description: Attendance marked
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAttendanceBody"
      responses:
        "200":
          description: Updated attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/summary:
    get:
      operationId: getAttendanceSummary
      tags: [attendance]
      summary: Get attendance summary for a month
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceSummary"
...
components:
  schemas:
...
    Attendance:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        date:
          type: string
        status:
          type: string
          description: present, absent, half_day
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        overtimeHours:
          type: ["number", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - employeeId
        - employeeName
        - date
        - status
        - createdAt
    MarkAttendanceBody:
      type: object
      properties:
        records:
          type: array
          items:
            type: object
            properties:
              employeeId:
                type: integer
              date:
                type: string
              status:
                type: string
              checkIn:
                type: ["string", "null"]
              checkOut:
                type: ["string", "null"]
              note:
                type: ["string", "null"]
              overtimeHours:
                type: ["number", "null"]
              latitude:
                type: ["number", "null"]
              longitude:
                type: ["number", "null"]
            required:
              - employeeId
              - date
              - status
      required:
        - records
    UpdateAttendanceBody:
      type: object
      properties:
        status:
          type: string
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        overtimeHours:
          type: ["number", "null"]
    AttendanceSummary:
      type: object
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        present:
          type: integer
        absent:
          type: integer
        halfDay:
          type: integer
        overtime:
          type: number
        workingDays:
          type: integer
        percentage:
          type: number
      required:
        - employeeId
        - employeeName
        - present
        - absent
        - halfDay
        - overtime
        - workingDays
        - percentage
...